28/05/2020 Finding remaining element in array
Given an array arr[] of size N and an integer K. The task is to find the last remaining element in the array after reducing the array. The rules for reducing the array are: The first and last element say X and Y are chosen and removed from the array arr[]. The values X and Y are added. Z = X + Y. Insert the value of Z % K into the array arr[] at the position ((N/2) + 1)th position, where N denotes the current length of the array.